pyproject.toml
[tool.poetry]
name = "cashbox-client"
version = "0.1.0"
description = ""
authors = ["Kostya Lukyanchikov <lukyanchikovk@gmail.com>"]
readme = "README.md"
packages = [ {include = "src"}, ]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyside6 = "^6.7.2"
websocket-client = "^1.8.0"
darkdetect = "^0.8.0"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.10.0"
ruff = "^0.4.4"
black = "^24.4.2"
mypy = "^1.10.0"
bandit = "^1.7.8"
pyupgrade-directories = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
build-installer = "build:full_build"

[tool.mypy]
exclude = "lib"
strict = true
python_version = "3.11"
ignore_missing_imports = true
install_types = true
non_interactive = true

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
fix = true
line-length = 88
target-version = "py311"
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "Q", # flake8-quotes
]
lint.ignore = [
    "E501", # line too long, handled by black
    "C901", # too complex
]
lint.pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"connector/application/dependencies.py" = ["B008"]
"connector/application/subscribers.py" = ["B008"]
